refactor(hunt): extract found-log + zap ingest into useFoundLogIngest; fix stale-flush race
HuntPiggyDetailScreen.tsx grew past its file-size baseline (1710 lines) —
extract the found-log and zap-receipt relay ingest/batching logic this PR
added into src/hooks/useFoundLogIngest.ts. The hook owns both coalesced
subscriptions, their derived sortedLogs/zapTotalsByLog views, and the
optimistic-insert used after posting a log; the screen shrinks to 1577
lines and just consumes the hook's return value.

Also fixes the unresolved Copilot thread: flushLogs cleared the shared
logFlushTimerRef before checking `cancelled`, so a stale/cancelled call
could null out a *newer* effect instance's already-scheduled timer (the
timer/pending refs are reused across effect re-runs on coord change),
silently stalling that instance's auto-flush until another event arrived.
Reordered so the cancelled check runs first and a stale call returns
before touching the shared timer ref or buffered state.
